requirements.txt
asgiref==3.3.1
astroid==2.4.2
Django==3.1.5
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-createsuperuserwithpassword==2.0.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
isort==5.7.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2-binary==2.8.6
PyJWT==2.0.1
pylint==2.6.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
